Migrate the delete/remove pages
